refac(witnessgen): add runner to executor and witnessgen
